requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
filelock==3.16.1
fsspec==2024.12.0
huggingface-hub==0.27.1
idna==3.10
packaging==24.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
